python/kfserving/requirements.txt
certifi>=14.05.14
six>=1.15
python_dateutil>=2.5.3
setuptools>=21.0.0
urllib3>=1.15.1
kubernetes>=12.0.0
tornado>=6.0.0
argparse>=1.4.0
minio>=4.0.9,<7.0.0
google-cloud-storage>=1.31.0
adal>=1.2.2
table_logger>=0.3.5
numpy>=1.17.3
azure-storage-blob>=12.7.1,<13.0.0
cloudevents>=1.2.0
avro>=1.10.1
